Engine: Move internal type system under Types namespace
Moved the internal type system classes from Spark.Search to
Spark.Engine.Search.Types.
