+++ src/Spark.Engine/Core/SearchResults.cs
@@ -1,15 +1,14 @@
-﻿/* 
+﻿/*
  * Copyright (c) 2014-2018, Firely <info@fire.ly>
  * Copyright (c) 2021-2025, Incendi <info@incendi.no>
- * 
+ *
  * SPDX-License-Identifier: BSD-3-Clause
  */
 
+using Hl7.Fhir.Model;
+using Spark.Engine.Search.Types;
 using System.Collections.Generic;
 using System.Linq;
-
-using Hl7.Fhir.Model;
-using Spark.Search;
 
 namespace Spark.Engine.Core;
