+++ src/Spark.Engine/Extensions/QuantityExtensions.cs
@@ -1,17 +1,17 @@
-﻿/* 
+﻿/*
  * Copyright (c) 2015-2018, Firely <info@fire.ly>
  * Copyright (c) 2018-2025, Incendi <info@incendi.no>
- * 
+ *
  * SPDX-License-Identifier: BSD-3-Clause
  */
 
 using Fhir.Metrics;
-using FM = Hl7.Fhir.Model;
+using Spark.Engine.Model;
+using Spark.Engine.Search.Types;
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using Spark.Search;
-using Spark.Engine.Model;
+using FM = Hl7.Fhir.Model;
 
 namespace Spark.Engine.Extensions;
