+++ src/Spark.Engine/Model/IndexEntry.cs
@@ -1,11 +1,11 @@
-﻿/* 
+﻿/*
  * Copyright (c) 2015-2018, Firely <info@fire.ly>
  * Copyright (c) 2021-2025, Incendi <info@incendi.no>
- * 
+ *
  * SPDX-License-Identifier: BSD-3-Clause
  */
 
-using Spark.Search;
+using Spark.Engine.Search.Types;
 using System.Collections.Generic;
 using System.Linq;
