+++ src/Spark.Engine/Search/ReferenceNormalizationService.cs
@@ -7,7 +7,7 @@
 using System;
 using System.Linq;
 using Spark.Engine.Core;
-using Spark.Search;
+using Spark.Engine.Search.Types;
 
 namespace Spark.Engine.Search;
