+++ src/Spark.Engine/Search/Types/DateValue.cs
@@ -9,7 +9,7 @@ using Hl7.Fhir.Model;
 using Spark.Search.Support;
 using System;
 
-namespace Spark.Search;
+namespace Spark.Engine.Search.Types;
 
 public class DateValue : ValueExpression
 {
