+++ src/Spark.Engine/Search/Types/Expression.cs
@@ -4,7 +4,7 @@
  * SPDX-License-Identifier: BSD-3-Clause
  */
 
-namespace Spark.Search;
+namespace Spark.Engine.Search.Types;
 
 public abstract class Expression
 {
