+++ src/Spark.Engine/Search/Types/QuantityValue.cs
@@ -9,7 +9,7 @@ using Hl7.Fhir.Serialization;
 using Spark.Search.Support;
 using System;
 
-namespace Spark.Search;
+namespace Spark.Engine.Search.Types;
 
 public class QuantityValue : ValueExpression
 {
