+++ src/Spark.Engine/Search/Types/StringValue.cs
@@ -5,7 +5,7 @@
  * SPDX-License-Identifier: BSD-3-Clause
  */
 
-namespace Spark.Search;
+namespace Spark.Engine.Search.Types;
 
 public class StringValue : ValueExpression
 {
