+++ src/Spark.Engine/Search/Types/UntypedValue.cs
@@ -6,7 +6,7 @@
 
 using Hl7.Fhir.Model;
 
-namespace Spark.Search;
+namespace Spark.Engine.Search.Types;
 
 public class UntypedValue : ValueExpression
 {
