+++ src/Spark.Engine/Search/Types/ValueExpression.cs
@@ -5,7 +5,7 @@
  * SPDX-License-Identifier: BSD-3-Clause
  */
 
-namespace Spark.Search;
+namespace Spark.Engine.Search.Types;
 
 public abstract class ValueExpression : Expression
 {
